requirements.txt
pyyaml
logging
watchdog
observer
requests
hiredis
rq
rratelimit
